Adjust type exports
