+++ packages/rpc-provider/src/substrate-connect/index.ts
@@ -392,6 +392,3 @@ export class ScProvider implements ProviderInterface {
     return this.send(method, [id]);
   }
 }
-
-export type ScProviderClass = typeof ScProvider;
-export type Config = SubstrateConnect.Config;
